add workflows seed fix
