+++ apps/workflow/server/prisma/example_scoop_portal_seed.js
@@ -1,160 +1,6 @@
-const { permissionTypes } = require("../api/consts");
-const { importMetadata } = require("../api/helpers/metadata");
 const { PrismaClient } = require("@prisma/client");
+const { createWorkflow } = require("./seedUtils");
 const prisma = new PrismaClient();
-
-/**
- * Create a workflow with the workflowData provided.
- *
- * @param {Object} workflowData
- * @returns An Object representing the workflow created
- */
-async function createWorkflow(workflowData) {
-  const baseActionData = {};
-
-  // required
-  baseActionData.name = workflowData.name;
-
-  // optional
-  if (workflowData.description)
-    baseActionData.description = workflowData.description;
-  if (workflowData.isFrozen) data.isFrozen = workflowData.isFrozen;
-  if (workflowData.metadata) {
-    baseActionData.metadata = {
-      create: importMetadata(workflowData.metadata),
-    };
-  }
-  if (workflowData.previousActionId) {
-    baseActionData.previousAction = {
-      connect: { id: workflowData.previousActionId },
-    };
-  }
-  if (workflowData.parentActionId) {
-    baseActionData.parentAction = {
-      connect: { id: workflowData.parentActionId },
-    };
-  }
-
-  let workflow;
-
-  await prisma.$transaction(async () => {
-    // Create the workflow
-    workflow = await prisma.workflowAttributes.create({
-      data: {
-        baseAction: {
-          create: {
-            ...baseActionData,
-            actionType: "workflow",
-            permissions: {
-              createMany: {
-                data: permissionTypes.map((permissionType) => ({
-                  userId: workflowData.userId,
-                  permissionType: permissionType,
-                })),
-              },
-            },
-          },
-        },
-      },
-    });
-
-    // Create any actions in this workflow
-    if (workflowData.actions?.length > 0) {
-      let lastActionId = null;
-
-      for (let i = 0; i < workflowData.actions.length; i++) {
-        const actionData = workflowData.actions[i];
-
-        lastActionId
-          ? (actionData.previousActionId = lastActionId)
-          : (actionData.rootActionOfId = workflow.id);
-
-        const action = await createAction(actionData);
-        lastActionId = action.id;
-      }
-    }
-  });
-
-  return workflow;
-}
-
-/**
- * Create an action with the actionData provided.
- *
- * @param {Object} actionData
- * @returns An Object representing the action created
- */
-async function createAction(actionData) {
-  // For workflow base actions
-  if (actionData.actionType && actionData.actionType === "workflow") {
-    // Create a workflow and return the base action as the newly create action
-    const workflow = await createWorkflow(actionData);
-    // throw Error(actionData.previousActionId);
-    return await prisma.action.findUnique({
-      where: { id: workflow.baseActionId },
-    });
-  }
-
-  // Get the data used to create the action
-  const data = {};
-
-  // required data
-  data.name = actionData.name;
-
-  // optional data
-  if (actionData.description) data.description = actionData.description;
-  if (actionData.form) data.form = actionData.form;
-  if (actionData.actionType) data.actionType = actionData.actionType;
-  if (actionData.isFrozen) data.isFrozen = actionData.isFrozen;
-  if (actionData.metadata) {
-    data.metadata = {
-      create: importMetadata(actionData.metadata),
-    };
-  }
-  // actions directly in complex/branching actions only
-  if (actionData.parentActionId) {
-    data.parentAction = { connect: { id: actionData.parentActionId } };
-  }
-  // actions directly in workflows only
-  if (actionData.previousActionId) {
-    data.previousAction = { connect: { id: actionData.previousActionId } };
-  }
-  if (actionData.rootActionOfId) {
-    data.rootActionOf = { connect: { id: actionData.rootActionOfId } };
-  }
-
-  let action;
-
-  await prisma.$transaction(async () => {
-    // Create the data on the database using prisma
-    action = await prisma.action.create({
-      data: data,
-    });
-
-    //make an action state for this actiony
-    await prisma.actionState.create({
-    data: {
-      stateType: 'notStarted',
-      actionId: action.id,
-      index: 0
-    },
-    });
-
-
-
-    // Create any child actions for this action
-    if (actionData.childActions?.length > 0) {
-      for (let i = 0; i < actionData.childActions.length; i++) {
-        const childAction = actionData.childActions[i];
-        childAction.parentActionId = action.id;
-
-        await createAction(childAction);
-      }
-    }
-  });
-
-  return action;
-}
 
 async function main() {
